backend/requirements.txt
async-timeout==4.0.2
click==8.1.3
Deprecated==1.2.13
environs==9.5.0
Flask==2.2.2
googleapis-common-protos==1.56.4
grpcio==1.49.0
grpcio-tools==1.49.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
packaging==21.3
protobuf==4.21.6
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-dotenv==0.21.0
redis==4.3.4
six==1.16.0
Werkzeug==2.2.2
wrapt==1.14.1
